project/persistence/backoffice-jdbc-dao/src/main/java/it/maxmin/plain/dao/config/SpringJdbcTemplateCfg.java
package it.maxmin.plain.dao.config;

import javax.sql.DataSource;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.Import;
import org.springframework.jdbc.core.JdbcTemplate;

@Import(JndiDataSourceCfg.class)
@Configuration
public class SpringJdbcTemplateCfg {

	private static Logger LOGGER = LoggerFactory.getLogger(SpringJdbcTemplateCfg.class);

	@Autowired
	DataSource dataSource;

	@Bean
	public JdbcTemplate jdbcTemplate() {
		JdbcTemplate jdbcTemplate = new JdbcTemplate();
		jdbcTemplate.setDataSource(dataSource);
		return jdbcTemplate;
	}
}
